style: move errdefiner as sub of error package
